backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.0
pydantic-settings==2.6.0
email-validator==2.2.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
bcrypt==3.2.2
boto3==1.35.20
httpx==0.27.0
python-multipart
apscheduler==3.10.4
